float64 изменён на int64

+++ backend/pkg/wallet/wallet.go
@@ -4,5 +4,5 @@ import "github.com/google/uuid"
 
 type Wallet struct {
 	ID     uuid.UUID
-	Amount float64
+	Amount int64
 }
